1.Change Integer to Long for response code and Pin length.

+++ app/src/main/java/com/pax/pay/poslink/ui/demo/base/RespStatusImpl.java
@@ -25,7 +25,7 @@ public class RespStatusImpl implements IRespStatus {
     }
 
     @Override
-    public void onDeclined(int code, String message) {
+    public void onDeclined(long code, String message) {
         Activity activity = activityWeakReference.get();
         if (activity != null) {
             activity.runOnUiThread(() -> {
